Make scope crawling one way
Rather than allowing lower scoped statements write access/ mutate rights to higher scoped values, error out at compilation. This is designed to ensure that globals are, in effect, constant which avoids stupid developers (hello, is me) from accidentally breaking things.

This is absolutely a feature, and not because I couldn't figure out how to do the alernative; I totally didn't manage to sell it to myself after as actually a good feature (which it is)