Ensure knowsWhere() reflects zero'd line/ col information
As per https://www.gnu.org/software/bison/manual/html_node/Location-Type.html, the fields we need are initialised as all 1s. Thus, we need to check for 1s to suss out whether we know the position of an error.

Note: this does mean an error at line 1, char 1 will not throw the correct error, but given that our grammar defines a file starting the the package keyword we'd get a syntax error way before this anyway